Add parsing/unparsing for Hexagon levels.
The format for the input and output is not the same for now. Might
want to address that in the future.

Note that solving Hexagon levels doesn't work yet. It turns out that
numpy is not good enough for this because it doesn't support solving
modulo m. More on this in a future patch.